+++ src/documentBuilder/InferResult/__.test-d.ts
@@ -1,7 +1,7 @@
 import type { Date } from '../../../tests/_/fixtures/scalars.js'
 import type { db } from '../../../tests/_/schemas/db.js'
-import type { Schema } from '../../../tests/_/schemas/kitchen-sink/graffle/modules2/schema.js'
-import type * as SelectionSets from '../../../tests/_/schemas/kitchen-sink/graffle/modules2/selection-sets.js'
+import type { Schema } from '../../../tests/_/schemas/kitchen-sink/graffle/modules/schema.js'
+import type * as SelectionSets from '../../../tests/_/schemas/kitchen-sink/graffle/modules/selection-sets.js'
 import { assertEqual } from '../../lib/assert-equal.js'
 import type { Registry } from '../../types/Schema/nodes/Scalar/helpers.js'
 import type { InferResult } from './__.js'
