+++ src/documentBuilder/SelectGraphQLMapper/toGraphQL.test.ts
@@ -1,8 +1,8 @@
 import { expect, test } from 'vitest'
 import { Date } from '../../../tests/_/fixtures/scalars.js'
 import { db } from '../../../tests/_/schemas/db.js'
-import { schemaDrivenDataMap } from '../../../tests/_/schemas/kitchen-sink/graffle/modules2/schema-driven-data-map.js'
-import type * as SelectionSets from '../../../tests/_/schemas/kitchen-sink/graffle/modules2/selection-sets.js'
+import { schemaDrivenDataMap } from '../../../tests/_/schemas/kitchen-sink/graffle/modules/schema-driven-data-map.js'
+import type * as SelectionSets from '../../../tests/_/schemas/kitchen-sink/graffle/modules/selection-sets.js'
 import { Grafaid } from '../../lib/grafaid/__.js'
 import type { Schema } from '../../types/Schema/__.js'
 import { Select } from '../Select/__.js'
