+++ src/extensions/CustomScalars/encode.test.ts
@@ -3,7 +3,7 @@ import { Date } from '../../../tests/_/fixtures/scalars.js'
 import { test } from '../../../tests/_/helpers.js'
 import { db } from '../../../tests/_/schemas/db.js'
 import type { Graffle } from '../../../tests/_/schemas/kitchen-sink/graffle/__.js'
-import { schemaDrivenDataMap } from '../../../tests/_/schemas/kitchen-sink/graffle/modules2/schema-driven-data-map.js'
+import { schemaDrivenDataMap } from '../../../tests/_/schemas/kitchen-sink/graffle/modules/schema-driven-data-map.js'
 import { Spy } from '../../../tests/_/SpyExtension.js'
 import { Select } from '../../documentBuilder/Select/__.js'
 import { SelectionSetGraphqlMapper } from '../../documentBuilder/SelectGraphQLMapper/__.js'
