+++ tests/_/schemas/kitchen-sink/graffle/modules/global.ts
