+++ tests/_/schemas/kitchen-sink/graffle/modules/methods-document.ts
