+++ tests/_/schemas/kitchen-sink/graffle/modules/methods-root.ts
