+++ tests/_/schemas/kitchen-sink/graffle/modules/methods-select.ts
