+++ tests/_/schemas/kitchen-sink/graffle/modules/scalar.ts
