+++ tests/_/schemas/kitchen-sink/graffle/modules/schema-driven-data-map.ts
