+++ tests/_/schemas/query-only/graffle/modules/methods-document.ts
