+++ tests/_/schemas/query-only/graffle/modules/scalar.ts
