+++ tests/_/schemas/query-only/graffle/modules/schema-driven-data-map.ts
